Fix ordering cycles starting with multi-user.target
This change introduces a new greenboot-success.target which requires
boot-complete.target and multi-user.target.
This is to avoid an ordering cycle that was introduced with the recent
addition of the service monitor, which explicitly orders
multi-user.target before boot-complete.target.
The start of services that were either WantedBy= or
RequiredBy=multi-user.target but which also had to run
After=boot-complete.target, couldn't be ordered properly
anymore. This change fixes that.

An unfortunate side-effect of this is that `multi-user.target` can't be
used as the default target anymore. The default target now has to be
changed to `greenboot-success.target` for greenboot to run.