support for read only /boot mount
Support is added for greenboot to work with read-only /boot.
It uses PrivateMounts=yes to isolate the system and remounts
boot as rw followed by setting the grub parameters as per healthcheck.

The following services are affected:

greenboot-grub2-set-success.service
greenboot-grub2-set-counter.service
greenboot-rpm-ostree-check-fallback.service

The grub2-editenv command is decoupled from greenboot-grub2-set-success.service
and is part of a separate script.

Signed-off-by: saypaul <paul.sayan@gmail.com>