rollback: if no bootc, fallback to rpm-ostree
Newer rpm-ostree in 9.6 pulls in bootc, which made our test fails
because we weren't rolling back the deployment.
This patch puts the `return` where it should be in case we're not
"bootc-native". That way we can safely fallback to rpm-ostree rollback
(bootc rollback should work tho...). A fallout of this nasty bug is that
we don't get any logs either so just adding some for mental sanity.

Signed-off-by: Antonio Murdaca <antoniomurdaca@gmail.com>